Restructure the directory and add Home page
