+++ src/views/Home.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Home() {
+  return (
+    <div>
+      <h1>Home Page</h1>
+    </div>
+  );
+}
+
+export default Home;
